requirements.txt
viam-sdk
pyinstaller
RPi.GPIO
adafruit-circuitpython-ht16k33
adafruit-extended-bus